Fix the Windows build: clamp the window position as int
RECT's members are LONG, so the middle of the area was a long while the
bounds were int, and std::clamp takes three of one type: MSVC found no
matching overload (C2672), which broke flutter build windows on the main
branch since "Ask to lock a folder again as soon as its window closes".

+++ windows/runner/flutter_window.cpp
@@ -28,14 +28,17 @@ void PlaceOver(HWND window, const RECT& area) {
     }
     const int width = bounds.right - bounds.left;
     const int height = bounds.bottom - bounds.top;
-    const int x = std::clamp((area.left + area.right - width) / 2,
-                             static_cast<int>(work.left),
-                             std::max(static_cast<int>(work.left),
-                                      static_cast<int>(work.right) - width));
-    const int y = std::clamp((area.top + area.bottom - height) / 2,
-                             static_cast<int>(work.top),
-                             std::max(static_cast<int>(work.top),
-                                      static_cast<int>(work.bottom) - height));
+    // All three of the same type: RECT's members are LONG.
+    const int x = std::clamp(
+        static_cast<int>((area.left + area.right - width) / 2),
+        static_cast<int>(work.left),
+        std::max(static_cast<int>(work.left),
+                 static_cast<int>(work.right) - width));
+    const int y = std::clamp(
+        static_cast<int>((area.top + area.bottom - height) / 2),
+        static_cast<int>(work.top),
+        std::max(static_cast<int>(work.top),
+                 static_cast<int>(work.bottom) - height));
     if (x == bounds.left && y == bounds.top) {
       return;
     }
